requirements.txt
api-client
cx_Freeze
python-dotenv
requests
reportlab